Dpm cli as a module
