+++ oceanum/cli/dpm/TODO.md
@@ -1,0 +1,51 @@
+
+# TODO list
+
+## Commands
+
+### Project
+
+- [ ] deploy [file]
+- [ ] validate [file]
+
+- [ ] list projects
+- [ ] delete project [file] | [project_name]
+
+#### Service Commands
+
+- [ ] list services
+- [ ] describe service [service_name]
+- [ ] delete project service [service_name] -p [project_name]
+- [ ] create project service [service_name] -p [project_name] --image [image_name] --port [port] --replicas [replicas] --env [env_vars] ...
+- [ ] update project service [service_name] -p [project_name]
+
+#### Stages Commands
+
+- [ ] list stages
+- [ ] describe stage [stage_name]
+- [ ] delete project stage [stage_name] -p [project_name]
+- [ ] create project stage [stage_name] -p [project_name]
+- [ ] update project stage [stage_name] -p [project_name]
+
+#### Pipeline Commands
+
+- [ ] list pipelines
+- [ ] describe pipeline [pipeline_name]
+- [ ] delete project pipeline [pipeline_name] -p [project_name]
+- [ ] create project pipeline [pipeline_name] -p [project_name]
+- [ ] update project pipeline [pipeline_name] -p [project_name]
+
+#### Tasks Commands
+
+- [ ] list tasks
+- [ ] describe task [task_name]
+- [ ] delete project task [task_name] -p [project_name]
+- [ ] create project task [task_name] -p [project_name]
+- [ ] update project task [task_name] -p [project_name]
+
+# Route Commands
+
+- [ ] list routes
+- [ ] describe route [route_name]
+
+
