+++ oceanum/cli/dpm/modelgen.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
-DPM_API_URL=${DPM_API_URL:-http://localhost:8000/api/}
 datamodel-codegen \
-    --url "${DPM_API_URL}openapi.json" \
+    --url "${1}openapi.json" \
     --input-file-type=openapi \
     --output=models.py \
     --target-python-version=3.11 \
